+++ src/mock/point.js
@@ -1,34 +1,11 @@
 import { getRandomElement, getRandomNumber, getRandomArray} from '../utils/utils';
-import { pointTypes, cities, offersSet } from './data';
-import { LoremIpsum } from 'lorem-ipsum';
+import { pointTypes, offersSet, destinations } from './data';
 import dayjs from 'dayjs';
 import { nanoid } from 'nanoid';
 
-
-const lorem = new LoremIpsum({
-  sentencesPerParagraph: {
-    max: 8,
-    min: 4
-  },
-  wordsPerSentence: {
-    max: 16,
-    min: 4
-  }
-});
-
 const generatePointType = () => getRandomElement(pointTypes);
 const generatePrice = () => getRandomNumber(100, 10000);
-const generatePictures = () => {
-  const isPictures = Boolean(getRandomNumber(0,1));
-  if (!isPictures) {
-    return null;
-  }
-  const generatePicture = () => ({
-    src:  `http://picsum.photos/300/200?r=${Math.random()}`,
-    description: lorem.generateSentences(1),
-  });
-  return Array.from({length: getRandomNumber(0, 5)}, generatePicture);
-};
+
 
 const generateDate = () => {
   const maxMinutGap = 6000;
@@ -44,6 +21,7 @@ const generateOffers = (type) => {
   }
 };
 
+
 export const generatePoint = () => {
   const date = generateDate();
   const offerType = generatePointType();
@@ -51,11 +29,7 @@ export const generatePoint = () => {
     basePrice: generatePrice(),
     dateFrom: date,
     dateTo: dayjs(date).add(getRandomNumber(3, 270) * 10, 'minute').toISOString(),
-    destination: {
-      description: getRandomNumber(0, 1) ? null : lorem.generateSentences(getRandomNumber(0, 5)),
-      name: getRandomElement(cities),
-      pictures: generatePictures(),
-    },
+    destination: destinations[getRandomNumber(0, destinations.length)],
     id: nanoid(),
     isFavorite: Boolean(getRandomNumber(0, 1)),
     type: offerType,
